Add files via upload
There can be many reasons, but the vast majority is because the program did something that the operating system did not like. Either some system function was called with invalid parameters, or the application tried to access resources it should not have access to the most common is an ‘out of bounds’ memory access, where the application is attempting to access memory out of its permitted ownership.